% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAds.R
\name{all_KijijiAds}
\alias{all_KijijiAds}
\title{Retrieve all Kijiji ads}
\usage{
all_KijijiAds(URL, exclude, pages = 3, outputFile = NULL)
}
\arguments{
\item{URL}{character, the address after doing a specific search on Kijiji}

\item{exclude}{vector, keywords to remove unwanted ads}

\item{pages}{numeric, total amount of pages to look for ads}

\item{outputFile}{character, name of the output file to save the retrieved ads.
If \code{NULL} it will return a list of ads}
}
\value{
list of ads
}
\description{
This function retrieves all ads from a Kijiji research URL. Third party ads
are removed.
}
\examples{
\dontrun{
# Retrieve all ads renting an apartment 3 1/2 in Montreal
# Exclude ads looking for an exchange
# And save the result in the 'ads.json' file
URL <- 'https://www.kijiji.ca/b-a-louer/ville-de-montreal/3-1-2/k0c30349001l1700281?price=__860'
all_KijijiAds(URL, exclude = 'echange', pages = 3, outputFile = 'ads.json')
}
}
