% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAds.R
\name{new_KijijiAds}
\alias{new_KijijiAds}
\title{Retrieve new Kijiji ads not present in the database}
\usage{
new_KijijiAds(
  URL,
  exclude = NULL,
  pages,
  inputFile,
  updateInput = TRUE,
  outputFile = NULL
)
}
\arguments{
\item{URL}{character, the address after doing a specific search on Kijiji}

\item{exclude}{vector, keywords to remove unwanted ads}

\item{pages}{numeric, total amount of pages to look for ads}

\item{inputFile}{character, name of the database file used in the \code{\link{all_KijijiAds}} function}

\item{updateInput}{logical, if FALSE it will not update the inputFile with the new retrieved ads.
Useful in case of debuging}

\item{outputFile}{character, name of the output file to save the retrieved new ads
If \code{NULL} it will retun a list of ads}
}
\value{
list of new ads
}
\description{
This function retrieves all new ads from a Kijiji search URL that are not present
in the defined database \code{inputFile}.
}
