% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{sendEmail}
\alias{sendEmail}
\title{Send email with new retrieved ads}
\usage{
sendEmail(
  mailInfo = "_mailInfo.yml",
  newAds = new_KijijiAds(URL, excludeWords, pages, inputFile)
)
}
\arguments{
\item{mailInfo}{character, the name of the file where all addresses and server
information are saved}

\item{newAds}{list, the output of the \code{\link{new_KijijiAds}} function}
}
\description{
This function is a wrapper of the emayili R package to send email messages
with the information of the new ads
}
\details{
The \code{mailInfo} file needs the following arguments: \cr
from: \cr
to: \cr
host: \cr
port: \cr
username: \cr
password: \cr
}
\examples{
\dontrun{
URL <- 'https://www.kijiji.ca/b-a-louer/ville-de-montreal/3-1-2/k0c30349001l1700281?price=__860'
sendEmail('_mailInfo.yml',
          newAds = new_KijijiAds(URL, excludeWords, pages = 2,
                                 inputFile = 'ads.json', updateInput = TRUE))
}
}
